requirements.txt
# Core API and server
fastapi==0.115.2
uvicorn[standard]==0.24.0

# RAG and Embeddings
groq==0.4.1
sentence-transformers==2.2.2
pinecone-client==2.2.4
faiss-cpu==1.7.4
huggingface_hub>=0.14.1  # compatible with sentence-transformers

# Web content loading and scraping
beautifulsoup4==4.12.2
requests==2.31.0
scrapy==2.11.0

# Data cache / environment
redis==5.0.1
python-dotenv==1.0.0

# NLP and tokenization
nltk==3.8.1
tiktoken==0.5.1

# Gradio UI
gradio==5.34.1

# Testing and Dev Tools
httpx==0.24.1
pytest==7.4.3
black==23.11.0
